fix(hooks): rebuild bundle and cover UserPromptSubmit malformed stdin
Address review feedback for the malformed stdin fail-open path.
