+++ tests/cli/stdin-reader.test.ts
@@ -1,29 +1,11 @@
 
 import { describe, it, expect, afterEach } from 'bun:test';
-import { Readable } from 'stream';
 
 import { readJsonFromStdin } from '../../src/cli/stdin-reader.js';
-
-const realStdin = process.stdin;
-const realStdinDescriptor = Object.getOwnPropertyDescriptor(process, 'stdin');
-
-function installFakeStdin(payload: string): void {
-  const fake = Readable.from([payload], { objectMode: false }) as unknown as NodeJS.ReadStream;
-  Object.defineProperty(fake, 'isTTY', { value: false, configurable: true });
-  Object.defineProperty(process, 'stdin', {
-    configurable: true,
-    enumerable: realStdinDescriptor?.enumerable ?? true,
-    writable: true,
-    value: fake,
-  });
-}
+import { installFakeStdin, restoreStdin } from '../fake-stdin.js';
 
 afterEach(() => {
-  if (realStdinDescriptor) {
-    Object.defineProperty(process, 'stdin', realStdinDescriptor);
-  } else {
-    Object.defineProperty(process, 'stdin', { value: realStdin, configurable: true, writable: true });
-  }
+  restoreStdin();
 });
 
 describe('readJsonFromStdin — onEnd contract (#2089)', () => {
